+++ app/src/androidTest/java/com/example/tsundokun/TsundokuDaoTest.kt
@@ -4,7 +4,7 @@ import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider.getApplicationContext
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.SmallTest
-import com.example.tsundokun.data.local.TsundokuEntity
+import com.example.tsundokun.data.local.entities.TsundokuEntity
 import com.example.tsundokun.data.local.TsundokunDatabase
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertNotNull
