+++ app/src/main/java/com/example/tsundokun/data/local/TsundokunDatabase.kt
@@ -2,9 +2,11 @@ package com.example.tsundokun.data.local
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.tsundokun.data.local.category.CategoryDao
-import com.example.tsundokun.data.local.category.CategoryEntity
-import com.example.tsundokun.data.local.tsundokucategory.TsundokuCategoryEntity
+import com.example.tsundokun.data.local.dao.CategoryDao
+import com.example.tsundokun.data.local.dao.TsundokuDao
+import com.example.tsundokun.data.local.entities.CategoryEntity
+import com.example.tsundokun.data.local.entities.TsundokuCategoryEntity
+import com.example.tsundokun.data.local.entities.TsundokuEntity
 
 @Database(entities = [TsundokuEntity::class, CategoryEntity::class, TsundokuCategoryEntity::class], version = 1, exportSchema = false)
 abstract class TsundokunDatabase : RoomDatabase() {
