+++ app/src/main/java/com/example/tsundokun/data/local/entities/CategoryEntity.kt
@@ -1,4 +1,4 @@
-package com.example.tsundokun.data.local.category
+package com.example.tsundokun.data.local.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
